Submit errors during validation process to serverError (see desc.)
These changes are directly sending the error objects to the
`serverError` response, which by default, should be safe and avoid
exposing data to the user in production mode.

With floatplane-hook-error, this is even better, as the Error object
will be logged out and wrapped into an “unknownError” that is sent to
the client. This data sent to the client will contain a unique ID which
can be used to correlate the error back to the log later on, if
necessary.

TLDR: This change and intended usage is secure. It’s for error
reporting.